move some functions that used to get data from jama to jama/tools
